+++ src/components/WebView.tsx
@@ -1,0 +1,96 @@
+import React, { useEffect, useMemo, useRef, useState } from 'react';
+import { Animated, Linking, Platform } from 'react-native';
+import { WebView as RnWebView, WebViewMessageEvent, WebViewNavigation } from 'react-native-webview';
+
+import { Error } from '~/components/Error';
+import theme from '~/styles/theme';
+
+interface WebViewProps {
+  uri: string;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  ref?: React.MutableRefObject<RnWebView<{}> | undefined>;
+  onMessage?: (event: WebViewMessageEvent) => void;
+  onNavigate?: (event: WebViewNavigation) => boolean;
+}
+
+export default function WebView({ uri, ref, onMessage, onNavigate }: WebViewProps) {
+  const [isError, setIsError] = useState(false);
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  const webViewRef = ref ?? useRef<RnWebView>();
+  const fadeAnimationRef = useRef(new Animated.Value(0));
+
+  const animationConfig: Animated.TimingAnimationConfig = useMemo(() => {
+    return { useNativeDriver: false, toValue: 1, duration: 1700 };
+  }, []);
+
+  useEffect(() => {
+    Animated.timing(fadeAnimationRef.current, animationConfig).start();
+  }, [animationConfig]);
+
+  const handleExternalLinks = (event: WebViewNavigation) => {
+    if (Platform.OS !== 'ios') {
+      return false;
+    }
+
+    const isExternalLink = event.navigationType === 'click';
+    if (isExternalLink) {
+      Linking.canOpenURL(event.url).then(supported => {
+        if (supported) {
+          Linking.openURL(event.url);
+        }
+      });
+      return false;
+    }
+    return true;
+  };
+
+  const handleNavigate = (event: WebViewNavigation) => {
+    if (onNavigate && onNavigate(event)) {
+      return true;
+    }
+    return handleExternalLinks(event);
+  };
+
+  if (isError) {
+    return (
+      <Error
+        reload={() => {
+          setIsError(false);
+          webViewRef.current?.reload();
+        }}
+      />
+    );
+  }
+
+  return (
+    <Animated.View
+      style={{
+        width: '100%',
+        height: '100%',
+        opacity: fadeAnimationRef.current,
+        backgroundColor: theme.color.background,
+      }}
+    >
+      <RnWebView
+        ref={ref => {
+          if (!ref) return;
+          webViewRef.current = ref;
+        }}
+        source={{ uri }}
+        bounces={false}
+        applicationNameForUserAgent={'YgtangApp/1.0'}
+        allowsBackForwardNavigationGestures
+        domStorageEnabled
+        onError={() => {
+          setIsError(true);
+        }}
+        onNavigationStateChange={handleNavigate}
+        onShouldStartLoadWithRequest={handleNavigate}
+        style={{
+          backgroundColor: theme.color.background,
+        }}
+        onMessage={onMessage}
+      />
+    </Animated.View>
+  );
+}
