+++ src/screens/HomeScreen.tsx
@@ -1,87 +1,18 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
-import { Animated, Linking, Platform } from 'react-native';
+import React from 'react';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { WebView, WebViewNavigation } from 'react-native-webview';
 
-import { Error } from '~/components/Error';
+import WebView from '~/components/WebView';
 import theme from '~/styles/theme';
 
 const uri = 'https://app.ygtang.kr/';
 
 export default function HomeScreen() {
-  const [isError, setIsError] = useState(false);
-  const webViewRef = useRef<WebView>();
-  const fadeAnimationRef = useRef(new Animated.Value(0));
-
-  const animationConfig: Animated.TimingAnimationConfig = useMemo(() => {
-    return { useNativeDriver: false, toValue: 1, duration: 1700 };
-  }, []);
-
-  useEffect(() => {
-    Animated.timing(fadeAnimationRef.current, animationConfig).start();
-  }, [animationConfig]);
-
-  const handleExternalLinks = (event: WebViewNavigation) => {
-    if (Platform.OS !== 'ios') {
-      return false;
-    }
-
-    const isExternalLink = event.navigationType === 'click';
-    if (isExternalLink) {
-      Linking.canOpenURL(event.url).then(supported => {
-        if (supported) {
-          Linking.openURL(event.url);
-        }
-      });
-      return false;
-    }
-    return true;
-  };
-
-  if (isError) {
-    return (
-      <Error
-        reload={() => {
-          setIsError(false);
-          webViewRef.current?.reload();
-        }}
-      />
-    );
-  }
-
   return (
     <SafeAreaView
       edges={['right', 'top', 'left']}
       style={{ flex: 1, backgroundColor: theme.color.background }}
     >
-      <Animated.View
-        style={{
-          width: '100%',
-          height: '100%',
-          opacity: fadeAnimationRef.current,
-          backgroundColor: theme.color.background,
-        }}
-      >
-        <WebView
-          ref={ref => {
-            if (!ref) return;
-            webViewRef.current = ref;
-          }}
-          source={{ uri }}
-          bounces={false}
-          applicationNameForUserAgent={'YgtangApp/1.0'}
-          allowsBackForwardNavigationGestures
-          domStorageEnabled
-          onError={() => {
-            setIsError(true);
-          }}
-          onNavigationStateChange={handleExternalLinks}
-          onShouldStartLoadWithRequest={handleExternalLinks}
-          style={{
-            backgroundColor: theme.color.background,
-          }}
-        />
-      </Animated.View>
+      <WebView uri={uri} />
     </SafeAreaView>
   );
 }
